Remove unicode listchars to fix #57
As discussed in Issue #57, setting Unicode characters for
`listchars` is a problematic default because of the scrolling
performance impact incurred when one of the specified glyphs is missing
from the user's selected font.

An alternative approach would be to choose "safer" unicode characters
that are likely to exist in most fonts. Given that the purpose of
vim-sensible is to provide universal sane defaults it seems this would
be best left to the user's `.vimrc` or another plugin.