Stop enabling shiftround; it breaks indent/outdent
The shiftround setting breaks many common cases of multi-line indent and
outdent for continuation lines. For example, suppose I have the
following code, in a buffer with 8-space indentation:

foo(x,
    y)

With the default settings, I can > those two lines, and end up with this:

        foo(x,
            y);

But with shiftround, I instead end up with:

        foo(x,
        y);

Thus, revert to the vim default of noshiftround.