Preserve fish as 'shell' for Vim 7.4.276 or later
As of 7.4.276, Vim understands how to run commands in subshells when
'shell' is fish.  This fixes the most common problems with running
external commands from Vim.

Plugins may still need to accommodate for fish, but fixing the cognitive
dissonance of having the user's shell changed outweighs the benefit to
lazy plugin writers.